\docType{package}
\name{RegCombinPack}
\alias{RegCombinPack}
\alias{RegCombinPack-package}
\title{RegCombinPack}
\description{
RegCombinPack is used to fit the regressions of all
combinations of a set of covaraites.
}
\examples{
\dontrun{
demo(fitRegCombin)
demo(performRegAnalysis)
}
}
\author{
Dalston G. Ward \email{ward.dalston@gmail.com}
}

