\docType{data}
\name{RegCombinPackData}
\alias{RegCombinPackData}
\title{Example Dataset}
\description{
This is some example data created using rpois and sample.
It can be used to run the functions in RegCombinPack
}
\details{
The variables included in the dataset are: \itemize{
\item\code{Y} A vector of 15 numbers sampled from 1:100
without replacement. \item\code{Variable 1} A vector of
numbers drawn from a Poisson distribution with lambda=26.
\item\code{Variable 2} A vector of numbers drawn from a
Poisson distribution with lambda=26. \item\code{Variable 3}
A vector of numbers drawn from a Poisson distribution with
lambda=26. \item\code{Variable 4} A vector of numbers drawn
from a Poisson distribution with lambda=26. }
}
\author{
Dalston G. Ward \email{ward.dalston@gmail.com}
}

